Addresed all comments Fixed some bugs
