+++ examples/convert.rs
@@ -157,9 +157,9 @@ fn main() {
 
             fs::write(&args[2], bytes.as_slice()).unwrap();
         }
-        #[cfg(feature = "glsl450-backend")]
+        #[cfg(feature = "glsl-out")]
         "vert" | "frag" => {
-            use naga::back::glsl450;
+            use naga::back::glsl;
 
             let mut file = fs::OpenOptions::new()
                 .write(true)
@@ -168,7 +168,7 @@ fn main() {
                 .open(&args[2])
                 .unwrap();
 
-            glsl450::write(&module, &mut file).unwrap();
+            glsl::write(&module, &mut file).unwrap();
         }
         other => {
             panic!("Unknown output extension: {}", other);
