+++ src/back/mod.rs
@@ -1,9 +1,7 @@
 //! Functions which export shader modules into binary and text formats.
 
-#[cfg(feature = "glsl450-backend")]
-pub mod glsl450;
-#[cfg(any(feature = "glsl450-backend"))]
-pub mod glsl_common;
+#[cfg(feature = "glsl-out")]
+pub mod glsl;
 pub mod msl;
 #[cfg(feature = "spirv")]
 pub mod spv;
